+++ mrb_doc/core/window.rb
@@ -7,15 +7,6 @@ def set_window_monitor(monitor)
   nil
 end
 
-# Gets the position of the specified monitor.
-# @param monitor [Integer]
-# @return [Vector2]
-def get_monitor_position(monitor)
-  # mrb_get_monitor_position
-  # src/mruby_integration/core/window.cpp
-  Vector2.new(0, 0)
-end
-
 # Gets the width of the specified monitor.
 # @param monitor [Integer]
 # @return [Integer]