+++ mrb_doc/models/monitor.rb
@@ -48,4 +48,24 @@ class Monitor
     # src/mruby_integration/models/monitor.cpp
     Monitor[0]
   end
+
+  # Returns the position of the {Monitor} as a {Vector2}.
+  #
+  # @example Basic usage
+  #   Window.open(
+  #     width: 1920,
+  #     height: 1080,
+  #     title: "My super cool game!"
+  #   )
+  #
+  #   puts Monitor.current.position
+  #   # => #<Vector2:0x558eed361650 x:0.0 y:0.0>
+  #
+  # @return [Vector2]
+  # @raise [Window::NotReadyError] If called before opening the Window
+  def position
+    # mrb_Monitor_position
+    # src/mruby_integration/models/monitor.cpp
+    Vector2[0, 0]
+  end
 end